rename *-util files to *-utils
closes #189
