+++ projects/ngx-d3timeline/src/lib/scale-utils.ts
